refactor: rewrite all CRUDs to match new siadal_schema

+++ Controllers/TeacherController.cs
@@ -25,7 +25,7 @@ namespace SIADAL.Controllers
         {
             if (!ModelState.IsValid) return BadRequest(ModelState);
 
-            try 
+            try
             {
                 var items = await _repository.GetAllAsync(query);
                 return Ok(items);
@@ -39,11 +39,11 @@ namespace SIADAL.Controllers
         [Authorize]
         [HttpGet("{id:int}")]
         [HttpGet("details/{id:int}")]
-        public async Task<IActionResult> Get(ulong id)
+        public async Task<IActionResult> Get(int id)
         {
             if (!ModelState.IsValid) return BadRequest(ModelState);
 
-            try 
+            try
             {
                 var item = await _repository.GetByIdAsync(id);
                 if (item == null) return NotFound();
@@ -76,11 +76,11 @@ namespace SIADAL.Controllers
         [Authorize]
         [HttpPut("{id:int}")]
         [HttpPut("update/{id:int}")]
-        public async Task<IActionResult> Put(ulong id, [FromBody] UpdateTeacherDTO dto)
+        public async Task<IActionResult> Put(int id, [FromBody] UpdateTeacherDTO dto)
         {
             if (!ModelState.IsValid) return BadRequest(ModelState);
 
-            try 
+            try
             {
                 var updatedItem = await _repository.UpdateAsync(id, dto);
                 if (updatedItem == null) return NotFound();
@@ -95,11 +95,11 @@ namespace SIADAL.Controllers
         [Authorize]
         [HttpDelete("{id:int}")]
         [HttpDelete("delete/{id:int}")]
-        public async Task<IActionResult> Delete(ulong id)
+        public async Task<IActionResult> Delete(int id)
         {
             if (!ModelState.IsValid) return BadRequest(ModelState);
 
-            try 
+            try
             {
                 var deleted = await _repository.DeleteAsync(id);
                 if (!deleted) return NotFound();
@@ -112,4 +112,3 @@ namespace SIADAL.Controllers
         }
     }
 }
-
